+++ CHANGELOG.md
@@ -15,7 +15,7 @@
 
 ## Changed
 
-* `plutus-core` version bound to `>=1.40.0.0 && <1.41`
+* `plutus-core` version bound to `>=1.45.0.0 && <1.46`
 * Now, default for constant hoisting is 20 bytes. This means any constant larger than 20 bytes in UPLC flat will get hoisted.
   This can be changed via compilation config.